requirements.txt
pandas==1.3.3
mysql-connector-python==8.0.26
pymongo==3.12.0
pytest==6.2.4
